Moved server-specific sources to their own folder. Removed some unused code as per Codacy reports
