+++ lib/api/entries/index.js
@@ -94,25 +94,6 @@ function configure (app, wares, ctx) {
       return next();
     }
 
-    // check for last modified from query results
-    
-    function ifModifiedSince (req, res, next) {
-    
-      var lastEntryDate = new Date(_.last(res.entries).dateString);
-      res.setHeader('Last-Modified', lastEntryDate.toUTCString());
-      
-      var ifModifiedSince = req.get('If-Modified-Since');
-
-      if (!ifModifiedSince) { return next(); }
-      
-      if (lastEntryDate <= new Date(ifModifiedSince)) {
-        res.status(304).send({status:304, message: 'Not modified', type:'internal'});
-        return;
-      }
-      
-      return next();
-    }
-
     /**
       * @method format_entries
       * A final middleware to send payloads assembled by previous middlewares
