+++ lib/authorization/storage.js
@@ -5,7 +5,7 @@ var crypto = require('crypto');
 var shiroTrie = require('shiro-trie');
 var ObjectID = require('mongodb').ObjectID;
 
-var find_options = require('../query');
+var find_options = require('../server/query');
 
 function init (env, ctx) {
   var storage = { };
