+++ lib/client/receiveddata.js
@@ -37,7 +37,6 @@ function mergeDataUpdate(isDelta, cachedDataArray, receivedDataArray, maxAge) {
   }
 
   // purge old data from cache before updating
-  var purgeCount = 0;
   var mAge = (isNaN(maxAge) || maxAge == null) ? TWO_DAYS : maxAge;
   var twoDaysAgo = new Date().getTime() - mAge;
     
@@ -45,14 +44,8 @@ function mergeDataUpdate(isDelta, cachedDataArray, receivedDataArray, maxAge) {
     var element = cachedDataArray[i];
       if (element !== null && element !== undefined && element.mills <= twoDaysAgo) {
         cachedDataArray.splice(i,0);
-        purgeCount += 1;
     }
   }
-  
-  //if (purgeCount > 0) {
-  //  console.log('Purged ' + purgeCount + ' elements from cache due to age');
-  //}
-
 
   // If this is delta, calculate the difference, merge and sort
   var diff = nsArrayDiff(cachedDataArray, receivedDataArray);
