+++ lib/client/renderer.js
@@ -393,8 +393,6 @@ function init (client, d3) {
 
   function calcTreatmentRadius(treatment, opts, carbratio) {
     var CR = treatment.CR || carbratio || 20;
-    var carbs = treatment.carbs || CR;
-    var insulin = treatment.insulin || 1;
     var carbsOrInsulin = CR;
     if ( treatment.carbs ) {
         carbsOrInsulin = treatment.carbs;
