+++ lib/plugins/direction.js
@@ -1,7 +1,5 @@
 'use strict';
 
-var _ = require('lodash');
-
 function init() {
 
   var direction = {
