+++ lib/plugins/iob.js
@@ -244,14 +244,6 @@ function init(ctx) {
   };
 
   function alexaIOBIntentHandler (callback, slots, sbx) {
-    var preamble =
-        (slots && slots.pwd && slots.pwd.value) ?
-            translate('alexaPreamble3person', {
-                params: [
-                    slots.pwd.value.replace('\'s', '')
-                ]
-              }) :
-            translate('alexaPreamble');
 
     var message = translate('alexaIobIntent', {
       params: [
