+++ lib/plugins/loop.js
@@ -8,7 +8,6 @@ var levels = require('../levels');
 // var ALL_STATUS_FIELDS = ['status-symbol', 'status-label', 'iob', 'freq', 'rssi']; Unused variable
 
 function init(ctx) {
-  var timeago = require('./timeago')(ctx);
   var utils = require('../utils')(ctx);
 
   var loop = {
