+++ lib/report_plugins/glucosedistribution.js
@@ -414,12 +414,10 @@ glucosedistribution.report = function report_glucosedistribution(datastorage, so
 
     var timeInT1 = Math.round(100 * t1count / events).toFixed(1);
     var timeInT2 = Math.round(100 * t2count / events).toFixed(1);
-    var mac = (total / events).toFixed(1);
 
     var unitString = ' mg/dl';
 
     if (client.settings.units == 'mmol') {
-        mac = (total / events / 18.0).toFixed(2);
         TDC = TDC / 18.0;
         TDCHourly = TDCHourly / 18.0;
         unitString = ' mmol/L';
