+++ lib/report_plugins/profiles.js
@@ -14,7 +14,6 @@ module.exports = init;
 
 profiles.html = function html(client) {
   var translate = client.translate;
-  var a = translate('Database records');
   var ret =
       '<h2>' + translate('Profiles') + '</h2>'
     + '<br>' + translate('Database records') + '&nbsp'
