+++ lib/server/bootevent.js
@@ -59,7 +59,7 @@ function boot (env, language) {
 
     try {
       if (_.startsWith(env.storageURI, 'openaps://')) {
-        require('./storage/openaps-storage')(env, function ready (err, store) {
+        require('../storage/openaps-storage')(env, function ready (err, store) {
           if (err) {
             throw err;
           }
@@ -70,7 +70,7 @@ function boot (env, language) {
         });
       } else {
         //TODO assume mongo for now, when there are more storage options add a lookup
-        require('./storage/mongo-storage')(env, function ready(err, store) {
+        require('../storage/mongo-storage')(env, function ready(err, store) {
           // FIXME, error is always null, if there is an error, the storage.js will throw an exception
           console.log('Mongo Storage system ready');
           ctx.store = store;
@@ -91,7 +91,7 @@ function boot (env, language) {
       return next();
     }
 
-    ctx.authorization = require('./authorization')(env, ctx);
+    ctx.authorization = require('../authorization')(env, ctx);
     ctx.authorization.storage.reload(function loaded (err) {
       if (err) {
         ctx.bootErrors = ctx.bootErrors || [ ];
@@ -109,13 +109,13 @@ function boot (env, language) {
     ///////////////////////////////////////////////////
     // api and json object variables
     ///////////////////////////////////////////////////
-    ctx.plugins = require('./plugins')({
+    ctx.plugins = require('../plugins')({
       settings: env.settings
       , language: ctx.language
     }).registerServerDefaults();
 
-    ctx.pushover = require('./plugins/pushover')(env);
-    ctx.maker = require('./plugins/maker')(env);
+    ctx.pushover = require('../plugins/pushover')(env);
+    ctx.maker = require('../plugins/maker')(env);
     ctx.pushnotify = require('./pushnotify')(env, ctx);
 
     ctx.entries = require('./entries')(env, ctx);
@@ -124,14 +124,14 @@ function boot (env, language) {
     ctx.profile = require('./profile')(env.profile_collection, ctx);
     ctx.food = require('./food')(env, ctx);
     ctx.pebble = require('./pebble')(env, ctx);
-    ctx.properties = require('./api/properties')(env, ctx);
-    ctx.bus = require('./bus')(env.settings, ctx);
-    ctx.ddata = require('./data/ddata')();
-    ctx.dataloader = require('./data/dataloader')(env, ctx);
-    ctx.notifications = require('./notifications')(env, ctx);
+    ctx.properties = require('../api/properties')(env, ctx);
+    ctx.bus = require('../bus')(env.settings, ctx);
+    ctx.ddata = require('../data/ddata')();
+    ctx.dataloader = require('../data/dataloader')(env, ctx);
+    ctx.notifications = require('../notifications')(env, ctx);
 
     if (env.settings.isEnabled('alexa')) {
-      ctx.alexa = require('./plugins/alexa')(env, ctx);
+      ctx.alexa = require('../plugins/alexa')(env, ctx);
     }
 
     next( );
@@ -174,7 +174,7 @@ function boot (env, language) {
     });
 
     ctx.bus.on('data-loaded', function updatePlugins ( ) {
-      var sbx = require('./sandbox')().serverInit(env, ctx);
+      var sbx = require('../sandbox')().serverInit(env, ctx);
       ctx.plugins.setProperties(sbx);
       ctx.notifications.initRequests();
       ctx.plugins.checkNotifications(sbx);
@@ -192,7 +192,7 @@ function boot (env, language) {
       return next();
     }
 
-    ctx.bridge = require('./plugins/bridge')(env);
+    ctx.bridge = require('../plugins/bridge')(env);
     if (ctx.bridge) {
       ctx.bridge.startEngine(ctx.entries);
     }
@@ -204,7 +204,7 @@ function boot (env, language) {
       return next();
     }
 
-    ctx.mmconnect = require('./plugins/mmconnect').init(env, ctx.entries, ctx.devicestatus);
+    ctx.mmconnect = require('../plugins/mmconnect').init(env, ctx.entries, ctx.devicestatus);
     if (ctx.mmconnect) {
       ctx.mmconnect.run();
     }
