+++ lib/server/devicestatus.js
