+++ lib/server/pebble.js
@@ -2,8 +2,8 @@
 
 var _ = require('lodash');
 
-var sandbox = require('./sandbox')();
-var units = require('./units')();
+var sandbox = require('../sandbox')();
+var units = require('../units')();
 
 var DIRECTIONS = {
   NONE: 0
@@ -169,7 +169,7 @@ function pebble (req, res) {
 }
 
 function configure (env, ctx) {
-  var wares = require('./middleware/')(env);
+  var wares = require('../middleware/')(env);
   function middle (req, res, next) {
     req.env = env;
     req.ctx = ctx;
