+++ lib/server/pushnotify.js
@@ -4,8 +4,8 @@ var _ = require('lodash');
 var crypto = require('crypto');
 var NodeCache = require('node-cache');
 
-var levels = require('./levels');
-var times = require('./times');
+var levels = require('../levels');
+var times = require('../times');
 
 function init(env, ctx) {
