+++ lib/server/websocket.js
@@ -1,10 +1,9 @@
 'use strict';
 
-var levels = require('./levels');
-var times = require('./times');
-var calcData = require('./data/calcdelta');
+var levels = require('../levels');
+var times = require('../times');
+var calcData = require('../data/calcdelta');
 var ObjectID = require('mongodb').ObjectID;
-var profilefunctions = require('./profilefunctions')();
        
 function init (env, ctx, server) {
