+++ server.js
@@ -45,7 +45,7 @@ function create (app) {
   return transport.createServer(app);
 }
 
-require('./lib/bootevent')(env, language).boot(function booted (ctx) {
+require('./lib/server/bootevent')(env, language).boot(function booted (ctx) {
     var app = require('./app')(env, ctx);
     var server = create(app).listen(PORT, HOSTNAME);
     console.log(translate('Listening on port'), PORT, HOSTNAME);
@@ -55,7 +55,7 @@ require('./lib/bootevent')(env, language).boot(function booted (ctx) {
     }
 
     if (env.MQTT_MONITOR) {
-      ctx.mqtt = require('./lib/mqtt')(env, ctx);
+      ctx.mqtt = require('./lib/server/mqtt')(env, ctx);
       var es = require('event-stream');
       es.pipeline(ctx.mqtt.entries, ctx.entries.map( ), ctx.mqtt.every(ctx.entries));
     }
@@ -63,7 +63,7 @@ require('./lib/bootevent')(env, language).boot(function booted (ctx) {
     ///////////////////////////////////////////////////
     // setup socket io for data and message transmission
     ///////////////////////////////////////////////////
-    var websocket = require('./lib/websocket')(env, ctx, server);
+    var websocket = require('./lib/server/websocket')(env, ctx, server);
 
     ctx.bus.on('data-processed', function() {
       websocket.update();
